requirements.txt
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.17.2
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
rich==13.7.1
six==1.16.0
spotipy==2.23.0
urllib3==2.2.1
